Change PlatformConstants.web to default
