+++ src/PlatformConstants.native.ts
@@ -1,0 +1,8 @@
+import { NativeModules, Platform } from './ReactCompact';
+
+type PlatformConstants = {
+  forceTouchAvailable: boolean;
+};
+
+export default (NativeModules?.PlatformConstants ??
+  Platform.constants) as PlatformConstants;
