--- src/PlatformConstants.web.ts
@@ -1,5 +1,0 @@
-export default {
-  get forceTouchAvailable() {
-    return false;
-  },
-};
